Won't echo for remote.
MUD clients shouldn't need to echo for the mud server and doing so would
require addition complexity for avoiding parsing by other protocols and
the like. Considering it shouldn't be a needed feature I'll leave this
for now. If it turns out it is needed, then it should be fixable.